networks: Ignore PCI `slot_path` of wireless devices
This is a stopgap measure to make more recent releases of TurrisHW
(>=0.11.0) work with current versions of foris-controller.

While TurrisHW already provides additional data for wireless devices,
namely PCI `slot_path`, processing of such additional data is not
implemented yet in foris-controller.

And it is not compatible with json schema of "networks" module either.

See issue:
https://gitlab.nic.cz/turris/foris-controller/foris-controller/-/issues/252